// shiki/types.ts
export interface IShikiTheme extends IRawTheme {
  /**
   * @description theme name
   */
  name: string;

  /**
   * @description light/dark theme
   */
  type: "light" | "dark" | "css";

  /**
   * @description tokenColors of the theme file
   */
  settings: IRawThemeSetting[];

  /**
   * @description text default foreground color
   */
  fg: string;

  /**
   * @description text default background color
   */
  bg: string;

  /**
   * @description relative path of included theme
   */
  include?: string;

  /**
   * @description color map of the theme file
   */
  colors?: Record<string, string>;
}